backend/app/gemini-video-summarise/requirements.txt
google-generativeai
python-dotenv
tqdm 